Restructuring of files complete
