+++ ASPNetCoreWebAPITraining.sln
@@ -7,6 +7,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ASPNetCoreWebAPITraining", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{7612C1E9-11C8-46FC-915E-C18BE6633E63}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "Application\Application.csproj", "{8719A59A-58F5-4803-B958-BA43996CD68B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{F03C8A4F-37C8-467C-A12C-F4DD80017D91}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +28,13 @@ Global
 		{7612C1E9-11C8-46FC-915E-C18BE6633E63}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7612C1E9-11C8-46FC-915E-C18BE6633E63}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7612C1E9-11C8-46FC-915E-C18BE6633E63}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8719A59A-58F5-4803-B958-BA43996CD68B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8719A59A-58F5-4803-B958-BA43996CD68B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8719A59A-58F5-4803-B958-BA43996CD68B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8719A59A-58F5-4803-B958-BA43996CD68B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F03C8A4F-37C8-467C-A12C-F4DD80017D91}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F03C8A4F-37C8-467C-A12C-F4DD80017D91}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F03C8A4F-37C8-467C-A12C-F4DD80017D91}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F03C8A4F-37C8-467C-A12C-F4DD80017D91}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
