+++ Application/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace Application;
+
+public class Class1
+{
+
+}
