+++ Domain/Common/BaseEntity.cs
@@ -4,8 +4,8 @@ public abstract class BaseEntity
 {
     protected BaseEntity(int id)
     {
-        _id = id;
+        Id = id;
     }
 
-     private readonly int _id;
+     public int Id { get; private set;}
 }
