+++ Infrastructure/Databases/MySqlDbContext.cs
@@ -1,7 +1,8 @@
 using Domain.Common;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 
-namespace ASPNetCoreWebAPITraining.Database
+namespace Infrastructure.Databases
 {
     public class MySqlDbContext : BaseDbContext
     {
