+++ Infrastructure/Databases/SqlServerDbContext.cs
@@ -1,7 +1,8 @@
-using ASPNetCoreWebAPITraining.Models;
+using Domain.Common;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 
-namespace ASPNetCoreWebAPITraining.Database
+namespace Infrastructure.Databases
 {
     public class SqlServerDbContext : BaseDbContext
     {
